chore(tests): move shared store contract out of megaplan package
Resolve megaplan/tests/ shadow package by moving the shared store-contract
helpers into a non-collected tests/contract/ tree and atomically rewriting
the two importers. megaplan/agent/tests/ (vendored hermes-agent test tree)
is intentionally untouched.

- git mv megaplan/tests/store_contract.py tests/contract/store_contract.py
- tests/contract/__init__.py (empty)
- tests/test_db_store.py / tests/test_file_store.py: rewrite imports
- megaplan/tests/__init__.py removed; empty package directory deleted
- docs/sprint-12b-cleanup-gate-inventory.md: update path reference

run_store_contract AND run_arnold_adapter_contract remain importable from
the new path. tests/contract/store_contract.py is not pytest-collected
(name not prefixed test_).
