--- megaplan/tests/__init__.py
@@ -1,1 +1,0 @@
-"""Test helpers bundled with the megaplan package."""
